moved shared code into more appropriate positions
