+++ dev/devicelab/bin/tasks/platform_channels_benchmarks_ios.dart
@@ -2,10 +2,12 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:flutter_devicelab/framework/adb.dart' show DeviceOperatingSystem;
-import 'package:flutter_devicelab/framework/framework.dart';
-import 'platform_channels_benchmarks.dart' as platform_channels_benchmarks;
+import 'package:flutter_devicelab/framework/framework.dart' show task;
+import 'package:flutter_devicelab/framework/adb.dart'
+    show DeviceOperatingSystem;
+import 'package:flutter_devicelab/tasks/platform_channels_benchmarks.dart'
+    as platform_channels_benchmarks;
 
 Future<void> main() async {
-  task(platform_channels_benchmarks.runTask(DeviceOperatingSystem.ios));
+  await task(platform_channels_benchmarks.runTask(DeviceOperatingSystem.ios));
 }
