+++ dev/devicelab/lib/tasks/microbenchmarks.dart
@@ -3,13 +3,13 @@
 // found in the LICENSE file.
 
 import 'dart:async';
-import 'dart:convert';
 import 'dart:io';
 
 import 'package:flutter_devicelab/framework/adb.dart';
 import 'package:flutter_devicelab/framework/framework.dart';
 import 'package:flutter_devicelab/framework/task_result.dart';
 import 'package:flutter_devicelab/framework/utils.dart';
+import 'package:flutter_devicelab/microbenchmarks.dart';
 import 'package:path/path.dart' as path;
 
 /// Creates a device lab task that runs benchmarks in
@@ -42,6 +42,7 @@ TaskFunction createMicrobenchmarkTask() {
 
         return readJsonResults(flutterProcess);
       }
+
       return _run();
     }
 
@@ -57,99 +58,9 @@ TaskFunction createMicrobenchmarkTask() {
       ...await _runMicrobench('lib/language/sync_star_semantics_bench.dart'),
       ...await _runMicrobench('lib/foundation/all_elements_bench.dart'),
       ...await _runMicrobench('lib/foundation/change_notifier_bench.dart'),
- };
+    };
 
-    return TaskResult.success(allResults, benchmarkScoreKeys: allResults.keys.toList());
+    return TaskResult.success(allResults,
+        benchmarkScoreKeys: allResults.keys.toList());
   };
-}
-
-Future<Process> startFlutter({
-  List<String> options = const <String>[],
-  bool canFail = false,
-  Map<String, String> environment,
-}) {
-  final List<String> args = flutterCommandArgs('run', options);
-  return startProcess(path.join(flutterDirectory.path, 'bin', 'flutter'), args, environment: environment);
-}
-
-Future<Map<String, double>> readJsonResults(Process process) {
-  // IMPORTANT: keep these values in sync with dev/benchmarks/microbenchmarks/lib/common.dart
-  const String jsonStart = '================ RESULTS ================';
-  const String jsonEnd = '================ FORMATTED ==============';
-  const String jsonPrefix = ':::JSON:::';
-  bool jsonStarted = false;
-  final StringBuffer jsonBuf = StringBuffer();
-  final Completer<Map<String, double>> completer = Completer<Map<String, double>>();
-
-  final StreamSubscription<String> stderrSub = process.stderr
-      .transform<String>(const Utf8Decoder())
-      .transform<String>(const LineSplitter())
-      .listen((String line) {
-        stderr.writeln('[STDERR] $line');
-      });
-
-  bool processWasKilledIntentionally = false;
-  bool resultsHaveBeenParsed = false;
-  final StreamSubscription<String> stdoutSub = process.stdout
-      .transform<String>(const Utf8Decoder())
-      .transform<String>(const LineSplitter())
-      .listen((String line) async {
-    print(line);
-
-    if (line.contains(jsonStart)) {
-      jsonStarted = true;
-      return;
-    }
-
-    if (line.contains(jsonEnd)) {
-      final String jsonOutput = jsonBuf.toString();
-
-      // If we end up here and have already parsed the results, it suggests that
-      // we have received output from another test because our `flutter run`
-      // process did not terminate correctly.
-      // https://github.com/flutter/flutter/issues/19096#issuecomment-402756549
-      if (resultsHaveBeenParsed) {
-        throw 'Additional JSON was received after results has already been '
-              'processed. This suggests the `flutter run` process may have lived '
-              'past the end of our test and collected additional output from the '
-              'next test.\n\n'
-              'The JSON below contains all collected output, including both from '
-              'the original test and what followed.\n\n'
-              '$jsonOutput';
-      }
-
-      jsonStarted = false;
-      processWasKilledIntentionally = true;
-      resultsHaveBeenParsed = true;
-      // Sending a SIGINT/SIGTERM to the process here isn't reliable because [process] is
-      // the shell (flutter is a shell script) and doesn't pass the signal on.
-      // Sending a `q` is an instruction to quit using the console runner.
-      // See https://github.com/flutter/flutter/issues/19208
-      process.stdin.write('q');
-      await process.stdin.flush();
-      // Also send a kill signal in case the `q` above didn't work.
-      process.kill(ProcessSignal.sigint);
-      try {
-        completer.complete(Map<String, double>.from(json.decode(jsonOutput) as Map<String, dynamic>));
-      } catch (ex) {
-        completer.completeError('Decoding JSON failed ($ex). JSON string was: $jsonOutput');
-      }
-      return;
-    }
-
-    if (jsonStarted && line.contains(jsonPrefix))
-      jsonBuf.writeln(line.substring(line.indexOf(jsonPrefix) + jsonPrefix.length));
-  });
-
-  process.exitCode.then<void>((int code) async {
-    await Future.wait<void>(<Future<void>>[
-      stdoutSub.cancel(),
-      stderrSub.cancel(),
-    ]);
-    if (!processWasKilledIntentionally && code != 0) {
-      completer.completeError('flutter run failed: exit code=$code');
-    }
-  });
-
-  return completer.future;
 }
